+++ vulnerabilities/forms.py
@@ -10,8 +10,12 @@ from django import forms
 from django.core.exceptions import ValidationError
 
 from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Field
+from crispy_forms.layout import Fieldset
+from crispy_forms.layout import Layout
 
 from dje.forms import DataspacedModelForm
+from dje.forms import Group
 from product_portfolio.models import ProductPackage
 from vulnerabilities.models import VulnerabilityAnalysis
 
@@ -21,6 +25,15 @@ class VulnerabilityAnalysisForm(DataspacedModelForm):
         choices=VulnerabilityAnalysis.Response.choices,
         widget=forms.CheckboxSelectMultiple,
         required=False,
+    )
+    propagate_to_products = forms.MultipleChoiceField(
+        label="Propagate analysis to:",
+        widget=forms.CheckboxSelectMultiple,
+        required=False,
+        help_text=(
+            "The listed products share the same vulnerable package. "
+            "The analysis values will be applied to all selected products."
+        ),
     )
 
     class Meta:
@@ -32,15 +45,25 @@ class VulnerabilityAnalysisForm(DataspacedModelForm):
             "justification",
             "responses",
             "detail",
+            "propagate_to_products",
         ]
         widgets = {
             "product_package": forms.widgets.HiddenInput,
             "vulnerability": forms.widgets.HiddenInput,
-            "detail": forms.Textarea(attrs={"rows": 3}),
+            "detail": forms.Textarea(attrs={"rows": 2}),
         }
 
     def __init__(self, user, *args, **kwargs):
+        affected_products = kwargs.pop("affected_products", [])
         super().__init__(user, *args, **kwargs)
+
+        propagate_to_field = self.fields["propagate_to_products"]
+        # Note: the whole `product` object is set as the label to be fully used in the
+        # template rendering.
+        propagate_to_products_choices = [(product.uuid, product) for product in affected_products]
+        propagate_to_field.choices = propagate_to_products_choices
+        if not propagate_to_products_choices:
+            propagate_to_field.widget = forms.widgets.HiddenInput()
 
         responses_model_field = self._meta.model._meta.get_field("responses")
         self.fields["responses"].help_text = responses_model_field.help_text
@@ -53,11 +76,33 @@ class VulnerabilityAnalysisForm(DataspacedModelForm):
     def helper(self):
         helper = FormHelper()
         helper.form_method = "post"
-        helper.form_id = "product-vulnerability-analysis-form"
         helper.form_tag = False
         helper.modal_title = "Vulnerability analysis"
         helper.modal_id = "vulnerability-analysis-modal"
+        helper.layout = Layout(
+            Fieldset(
+                "",
+                "product_package",
+                "vulnerability",
+                Group("state", "justification"),
+                "responses",
+                "detail",
+                Field(
+                    "propagate_to_products",
+                    template="vulnerabilities/forms/widgets/propagate_to_table.html",
+                ),
+            ),
+        )
         return helper
+
+    def save(self, *args, **kwargs):
+        instance = super().save(*args, **kwargs)
+
+        if products := self.cleaned_data.get("propagate_to_products"):
+            for product_uuid in products:
+                instance.propagate(product_uuid, self.user)
+
+        return instance
 
     def clean(self):
         main_fields = ["state", "justification", "responses", "detail"]